backend/requirements.txt
# Django and core dependencies
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-gis==1.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3
django-storages==1.14.2

# Database
psycopg2-binary==2.9.9

# Celery for background tasks
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Authentication and security
djoser==2.2.0
PyJWT==2.8.0
cryptography==41.0.7

# HTTP requests
requests==2.31.0
urllib3==2.0.7

# Data processing and analysis
numpy==1.25.2
pandas==2.1.3
scipy==1.11.4

# Image processing
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# Development and testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==19.12.0

# Code quality
flake8==6.1.0
black==23.10.1
isort==5.12.0

# Documentation
drf-spectacular==0.26.5

# Monitoring and logging
sentry-sdk==1.38.0

# File handling
openpyxl==3.1.2

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Utilities
slugify==0.0.1
validators==0.22.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0